Add DeviceQueuesBuilder

+++ src/vk/instance.rs
@@ -141,7 +141,7 @@ impl PhysicalDevice {
             vkGetPhysicalDeviceQueueFamilyProperties(self.handle, count.as_mut_ptr(), families.as_mut_ptr());
             let families: Vec<QueueFamily> = families.into_iter()
                 .enumerate()
-                .map(|(i,v)| QueueFamily::new(i, v))
+                .map(|(i, v)| QueueFamily::new(i as u32, v))
                 .collect();
             Ok(families)
         }
@@ -160,18 +160,19 @@ impl PhysicalDevice {
     }
 }
 
+#[derive(Clone)]
 pub struct QueueFamily {
-    index: usize,
+    index: u32,
     property: VkQueueFamilyProperties,
 }
 
 impl QueueFamily {
-    pub fn new(index: usize, property: VkQueueFamilyProperties) -> Self {
+    pub fn new(index: u32, property: VkQueueFamilyProperties) -> Self {
         QueueFamily { index: index, property: property }
     }
 
     #[inline]
-    pub fn index(&self) -> usize {
+    pub fn index(&self) -> u32 {
         self.index
     }
